Crate pallet_starknet

A Substrate pallet implementation for Starknet, a decentralized, permissionless, and scalable zk-rollup for general-purpose smart contracts. See the Starknet documentation for more information. The code consists of the following sections:

Config: The trait Config is defined, which is used to configure the pallet by specifying the parameters and types on which it depends. The trait also includes associated types for RuntimeEvent, StateRoot, SystemHash, and TimestampProvider.

Hooks: The Hooks trait is implemented for the pallet, which includes methods to be executed during the block lifecycle: on_finalize, on_initialize, on_runtime_upgrade, and offchain_worker.

Storage: Several storage items are defined, including Pending, CurrentBlock, BlockHash, ContractClassHashes, ContractClasses, Nonces, StorageView, LastKnownEthBlock, and FeeTokenAddress. These storage items are used to store and manage data related to the Starknet pallet.

Genesis Configuration: The GenesisConfig struct is defined, which is used to set up the initial state of the pallet during genesis. The struct includes fields for contracts, contract_classes, storage, fee_token_address, and _phantom. A GenesisBuild implementation is provided to build the initial state during genesis.

Events: A set of events are defined in the Event enum, including KeepStarknetStrange, StarknetEvent, and FeeTokenAddressChanged. These events are emitted during the execution of various pallet functions.

6.Errors: A set of custom errors are defined in the Error enum, which is used to represent various error conditions during the execution of the pallet.

Dispatchable Functions: The Pallet struct implements several dispatchable functions (ping, invoke, …), which allow users to interact with the pallet and invoke state changes. These functions are annotated with weight and return a DispatchResult.

Re-exports

pub use pallet::*;

pub use self::pallet::*;

Modules

message

The implementation of the message type.

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

runtime_api

The Starknet pallet’s runtime API

state_root

State root logic. Starknet state root logic.

transaction_validation

Transaction validation logic. Transaction validation logic.

types

The Starknet pallet’s runtime custom types. Starknet pallet custom types.

Macros

log

Constants

ETHEREUM_CONSENSUS_RPC

ETHEREUM_EXECUTION_RPC

SEQUENCER_ADDRESS