Module pallet_starknet::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub extern crate alloc;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Structs

GenesisConfig

Starknet genesis configuration.

Pallet

The pallet implementing the on-chain logic.

StarknetFee

Enums

Call

The Starknet pallet external functions. Dispatchable functions allows users to interact with the pallet and invoke state changes. These functions materialize as “extrinsics”, which are often compared to transactions. Dispatchable functions must be annotated with a weight and must return a DispatchResult.

Error

The Starknet pallet custom errors. ERRORS

Event

The Starknet pallet events. EVENTS See: <https://docs.substrate.io/main-docs/build/events-errors/>

Traits

Config

Configure the pallet by specifying the parameters and types on which it depends. We’re coupling the starknet pallet to the tx payment pallet to be able to override the fee mechanism and comply with starknet which uses an ER20 as fee token

Type Definitions

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.