Trait pallet_starknet::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Randomness: Randomness<Self::Hash, Self::BlockNumber>;
    type StateRoot: Get<U256>;
    type SystemHash: Hasher;
}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Randomness: Randomness<Self::Hash, Self::BlockNumber>

The type of Randomness we want to specify for this pallet.

type StateRoot: Get<U256>

How Starknet state root is calculated.

type SystemHash: Hasher

The hashing function to use.