Trait pallet_starknet::runtime_api::StarknetRuntimeApi

pub trait StarknetRuntimeApi<Block: BlockT>: Core<Block> {

    // Provided methods
    fn nonce(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        contract_address: ContractAddressWrapper
    ) -> Result<NonceWrapper, ApiError> { ... }
    fn nonce_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        contract_address: ContractAddressWrapper
    ) -> Result<NonceWrapper, ApiError> { ... }
    fn events(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Vec<EventWrapper>, ApiError> { ... }
    fn events_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<Vec<EventWrapper>, ApiError> { ... }
    fn get_storage_at(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        address: ContractAddressWrapper,
        key: StorageKeyWrapper
    ) -> Result<Result<Felt252Wrapper, DispatchError>, ApiError> { ... }
    fn get_storage_at_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        address: ContractAddressWrapper,
        key: StorageKeyWrapper
    ) -> Result<Result<Felt252Wrapper, DispatchError>, ApiError> { ... }
    fn call(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        address: ContractAddressWrapper,
        function_selector: Felt252Wrapper,
        calldata: Vec<Felt252Wrapper>
    ) -> Result<Result<Vec<Felt252Wrapper>, DispatchError>, ApiError> { ... }
    fn call_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        address: ContractAddressWrapper,
        function_selector: Felt252Wrapper,
        calldata: Vec<Felt252Wrapper>
    ) -> Result<Result<Vec<Felt252Wrapper>, DispatchError>, ApiError> { ... }
    fn contract_class_hash_by_address(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        address: ContractAddressWrapper
    ) -> Result<Option<ClassHashWrapper>, ApiError> { ... }
    fn contract_class_hash_by_address_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        address: ContractAddressWrapper
    ) -> Result<Option<ClassHashWrapper>, ApiError> { ... }
    fn contract_class_by_class_hash(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        class_hash: ClassHashWrapper
    ) -> Result<Option<ContractClassWrapper>, ApiError> { ... }
    fn contract_class_by_class_hash_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        class_hash: ClassHashWrapper
    ) -> Result<Option<ContractClassWrapper>, ApiError> { ... }
    fn chain_id(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Felt252Wrapper, ApiError> { ... }
    fn chain_id_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<Felt252Wrapper, ApiError> { ... }
    fn estimate_fee(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        transaction: Transaction
    ) -> Result<Result<(u64, u64), DispatchError>, ApiError> { ... }
    fn estimate_fee_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        transaction: Transaction
    ) -> Result<Result<(u64, u64), DispatchError>, ApiError> { ... }
    fn get_hasher(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Hasher, ApiError> { ... }
    fn get_hasher_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<Hasher, ApiError> { ... }
    fn extrinsic_filter(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        xts: Vec<<Block as BlockT>::Extrinsic>
    ) -> Result<Vec<Transaction>, ApiError> { ... }
    fn extrinsic_filter_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        xts: Vec<<Block as BlockT>::Extrinsic>
    ) -> Result<Vec<Transaction>, ApiError> { ... }

}

Provided Methods

fn nonce( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, contract_address: ContractAddressWrapper ) -> Result<NonceWrapper, ApiError>

Returns the nonce associated with the given address in the given block

fn nonce_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, contract_address: ContractAddressWrapper ) -> Result<NonceWrapper, ApiError>

Returns the nonce associated with the given address in the given block

fn events( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Vec<EventWrapper>, ApiError>

Returns the events associated with the given block

fn events_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<Vec<EventWrapper>, ApiError>

Returns the events associated with the given block

fn get_storage_at( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, address: ContractAddressWrapper, key: StorageKeyWrapper ) -> Result<Result<Felt252Wrapper, DispatchError>, ApiError>

Returns a storage slot value

fn get_storage_at_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, address: ContractAddressWrapper, key: StorageKeyWrapper ) -> Result<Result<Felt252Wrapper, DispatchError>, ApiError>

Returns a storage slot value

fn call( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, address: ContractAddressWrapper, function_selector: Felt252Wrapper, calldata: Vec<Felt252Wrapper> ) -> Result<Result<Vec<Felt252Wrapper>, DispatchError>, ApiError>

Returns a Call response.

fn call_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, address: ContractAddressWrapper, function_selector: Felt252Wrapper, calldata: Vec<Felt252Wrapper> ) -> Result<Result<Vec<Felt252Wrapper>, DispatchError>, ApiError>

Returns a Call response.

fn contract_class_hash_by_address( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, address: ContractAddressWrapper ) -> Result<Option<ClassHashWrapper>, ApiError>

Returns the contract class hash at the given address.

fn contract_class_hash_by_address_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, address: ContractAddressWrapper ) -> Result<Option<ClassHashWrapper>, ApiError>

Returns the contract class hash at the given address.

fn contract_class_by_class_hash( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, class_hash: ClassHashWrapper ) -> Result<Option<ContractClassWrapper>, ApiError>

Returns the contract class for the given class hash.

fn contract_class_by_class_hash_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, class_hash: ClassHashWrapper ) -> Result<Option<ContractClassWrapper>, ApiError>

Returns the contract class for the given class hash.

fn chain_id( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Felt252Wrapper, ApiError>

Returns the chain id.

fn chain_id_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<Felt252Wrapper, ApiError>

Returns the chain id.

fn estimate_fee( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, transaction: Transaction ) -> Result<Result<(u64, u64), DispatchError>, ApiError>

Returns fee estimate

fn estimate_fee_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, transaction: Transaction ) -> Result<Result<(u64, u64), DispatchError>, ApiError>

Returns fee estimate

fn get_hasher( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Hasher, ApiError>

Returns the hasher used by the runtime.

fn get_hasher_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<Hasher, ApiError>

Returns the hasher used by the runtime.

fn extrinsic_filter( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, xts: Vec<<Block as BlockT>::Extrinsic> ) -> Result<Vec<Transaction>, ApiError>

Filters extrinsic transactions to return only Starknet transactions

To support runtime upgrades, the client must be unaware of the specific extrinsic details. To achieve this, the client uses an OpaqueExtrinsic type to represent and manipulate extrinsics. However, the client cannot decode and filter extrinsics due to this limitation. The solution is to offload decoding and filtering to the RuntimeApi in the runtime itself, accomplished through the extrinsic_filter method. This enables the client to operate seamlessly while abstracting the extrinsic complexity.

fn extrinsic_filter_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, xts: Vec<<Block as BlockT>::Extrinsic> ) -> Result<Vec<Transaction>, ApiError>

Filters extrinsic transactions to return only Starknet transactions

To support runtime upgrades, the client must be unaware of the specific extrinsic details. To achieve this, the client uses an OpaqueExtrinsic type to represent and manipulate extrinsics. However, the client cannot decode and filter extrinsics due to this limitation. The solution is to offload decoding and filtering to the RuntimeApi in the runtime itself, accomplished through the extrinsic_filter method. This enables the client to operate seamlessly while abstracting the extrinsic complexity.