Module pallet_starknet::types

The Starknet pallet’s runtime custom types. Starknet pallet custom types.

Structs

EthBlockNumber

Struct that represents the response fields that we need of the eth node for eth_getBlockByNumber.

EthLogs

Struct that represents an Ethereum event for a message sent to starknet.

Message

Inner struct for messages.

NumberRes

Inner struct for block number.

Enums

OffchainWorkerError

Error enum wrapper for offchain worker tasks.

RawOrigin

Representation of the origin of a Starknet transaction. For now, we still don’t know how to represent the origin of a Starknet transaction, given that Starknet has native account abstraction. For now, we just use a dummy origin. See: https://github.com/keep-starknet-strange/madara/issues/21

Type Definitions

ContractClassMapping

ContractStorageKeyWrapper

Contract Storage Key

MaxTransactionsPendingBlock

Make this configurable. Max transaction/block

NonceWrapper

TODO: Replace with a proper type for field element. Nonce of a Starknet transaction.

StarkFeltWrapper

Felt

StorageKey

Storage Key